% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gareg-class.R
\docType{class}
\name{gareg-class}
\alias{gareg-class}
\title{S4 Class Definition for `gareg`
#'}
\description{
S4 Class for Genetic Algorithm-Based Regression
}
\section{Slots}{

\describe{
\item{\code{call}}{The matched call that created the object.}

\item{\code{N}}{The sample size of the time series.}
}}

